Optimize NPB-FT evolve() by using localAccess
This eliminates runtime locality checks, which are expensive compared to the
operations being done. This results in a 3-4x speedup for this step of the
computation. For size D at 8 locales this takes us from 100s to 80s.

Previously this code was:

```chpl
forall ijk in Dom {
  V[ijk] *= Twiddle[ijk];
  W[ijk] = V[ijk];
}
```

and now it is:

```chpl
forall ijk in Dom {
  V.localAccess[ijk] *= Twiddle.localAccess[ijk];
  W.localAccess[ijk] = V.localAccess[ijk];
}
```

Ideally we would write this as:

```chpl
forall (v, t, w) in zip(V, Twiddle, W) {
  v *= t;
  w = v;
}
```

or just:

```chpl
V *= Twiddle;
W = V;
```

But those 2 perform poorly