Add 'use' of SysCTypes and CPtr
SysCTypes is no longer available in Chapel code by default, so I'm
adding a [private] use of SysCTypes here to keep this working on
Chapel master.  In addition, we anticipate that CPtr will not be
available by default in Chapel 1.21, so I'm pre-emptively adding
a use of it for similar reasons and to avoid needing to come back
to this later.