Switch from Travis-CI to github actions
Github actions was released late in 2019 and provides a built-in
mechanism to do CI testing with github. Switch to using it since it's
more integrated with github, provides a cleaner syntax for using docker
images, and in my experience has been a lot more stable and reliable
compared to Travis over the last year or two.